refactor(diff): move binary diff context to separate file and remove unused variables in diffcontext
