+++ GitOut/Features/Git/Diff/DiffContext.cs
@@ -1,4 +1,3 @@
-using System;
 using System.IO;
 using System.Threading.Tasks;
 using GitOut.Features.IO;
@@ -7,17 +6,13 @@ namespace GitOut.Features.Git.Diff
 {
     public class DiffContext
     {
-        private readonly IGitRepository repository;
-
         private DiffContext(
             FileInfo info,
-            IGitRepository repository,
             GitFileId? source,
             GitFileId? destination
         )
         {
             FileExtension = info.Extension;
-            this.repository = repository;
             SourceId = source;
             DestinationId = destination;
         }
@@ -25,18 +20,16 @@ namespace GitOut.Features.Git.Diff
         private DiffContext(
             FileInfo info,
             BinaryDiffResult result,
-            IGitRepository repository,
             GitFileId? source,
             GitFileId? destination
-        ) : this(info, repository, source, destination) => Blob = result;
+        ) : this(info, source, destination) => Blob = result;
 
         private DiffContext(
             FileInfo info,
             TextDiffResult result,
-            IGitRepository repository,
             GitFileId? source,
             GitFileId? destination
-        ) : this(info, repository, source, destination) => Text = result;
+        ) : this(info, source, destination) => Text = result;
 
         public string FileExtension { get; }
 
@@ -45,10 +38,6 @@ namespace GitOut.Features.Git.Diff
 
         public BinaryDiffResult? Blob { get; }
         public TextDiffResult? Text { get; }
-
-        public Task<Stream> GetSourceStreamAsync() => SourceId is null
-            ? throw new InvalidOperationException("Cannot get stream for empty source")
-            : repository.GetBlobStreamAsync(SourceId);
 
         public static async Task<DiffContext> DiffAsync(
             IGitRepository repository,
@@ -103,52 +92,14 @@ namespace GitOut.Features.Git.Diff
             ? new DiffContext(
                 info,
                 new BinaryDiffResult(result.Blob!.Stream, repository, sourceId),
-                repository,
                 sourceId,
                 destinationId
             )
             : new DiffContext(
                 info,
                 result.Text,
-                repository,
                 sourceId,
                 destinationId
             );
-    }
-
-    public class BinaryDiffResult
-    {
-        private readonly Stream destinationStream;
-        private readonly IGitRepository repository;
-        private readonly GitFileId? sourceId;
-
-        private Stream? sourceStream;
-
-        public BinaryDiffResult(Stream destinationStream, IGitRepository repository, GitFileId? sourceId)
-        {
-            this.destinationStream = destinationStream;
-            this.repository = repository;
-            this.sourceId = sourceId;
-        }
-
-        public Stream GetBaseStream()
-        {
-            destinationStream.Position = 0;
-            return destinationStream;
-        }
-
-        public async Task<Stream> GetSourceStreamAsync()
-        {
-            if (sourceStream is null)
-            {
-                if (sourceId is null)
-                {
-                    throw new InvalidOperationException("Cannot get stream for empty source");
-                }
-                sourceStream = await repository.GetBlobStreamAsync(sourceId);
-            }
-            sourceStream.Position = 0;
-            return sourceStream;
-        }
     }
 }
